Use the `gutenbergSimpleSiteUser` account
The `defaultUser` account uses the `e2eflowtest3` site, which is used for theme testing as well. This is problematic because themes change the behavior of the customizer and can make this test fail. More specifically, themes with more than 1 widget area have a slight different UI when you navigate to the customizer with the widgets pane in focus. Instead of opening the block editor by default, it shows each area as an accordion button, and that breaks the logic.

Someday, it might make sense to handle that scenario (of > 1 widgets area) in this test, but for the regression test goals we have now, it wouldn't be needed. We're instead going to use a site that we know has a theme with only 1 widget area, and that is not used to test themes.